refactor: extract shared preset-conversion module
Move duplicated preset-to-timeframe conversion logic from both
ui/src/presets.ts and src/generate-category-ui-song.ts into a new
shared/preset-conversion.ts module. Update tsconfig to include shared/.
